Move webview protcol provider to own file
